pyproject.toml
[project]
name = "enderman-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "bguevarta",email = "brandon.guevara@infotec.mx"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi (>=0.136.1,<0.137.0)",
    "pydantic (>=2.13.3,<3.0.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "llama-index-core (>=0.14.21,<0.15.0)",
    "llama-index-llms-openai (>=0.7.5,<0.8.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "llama-index-embeddings-openai (>=0.6.0,<0.7.0)",
    "tzdata (>=2024.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
